move sort diary method to MainActivity
