+++ app/src/main/java/com/claire/traveldiary/map/showdiary/ShowDiaryPresenter.java
@@ -6,6 +6,8 @@ import android.widget.TextView;
 
 import com.claire.traveldiary.data.Diary;
 
+import java.util.List;
+
 import static com.bumptech.glide.util.Preconditions.checkNotNull;
 
 public class ShowDiaryPresenter implements ShowDiaryContract.Presenter {
@@ -43,6 +45,7 @@ public class ShowDiaryPresenter implements ShowDiaryContract.Presenter {
     public void setFontType(TextView title, TextView date) {
         mShowDiaryView.setFontTypeUi(title, date);
     }
+
 
     @Override
     public void start() {}
